Add completion for C#, Go, Rust and JavaScript
We need to switch to ubuntu since this provides mono and cargo in its
repositories.

We also stick all the command-line arguments in the one place as opposed
to having some of them in the Dockerfile; we get more control this way
(without having to rebuild the image each time).